Simplify the rds template file
We no longer need to interpolate anything into this template. All values
are expected to come from a `variables.tf` file in the user's
namespace/resources folder, and the module name is hard-coded as "rds"
in the cli tool. So we can just hard-code it here instead, and let the
cli tool do a simple read and write to save whatever is in this file as
`resources/rds.tf`