Add modify instance permissions
This change gives the IAM user created with the RDS instance additional
permissions, to enable users to upgrade their own RDS instances with
minimal involvement from the CP team.

The upgrade process is:

* Change RDS instance to use the defeault parameter group for it's
  db_engine
* Tell the pipeline to upgrade the RDS instance via a normal env. repo.
  PR

`rds:ModifyDBInstance` permission is required to change the instance's
DBParameterGroup.

Permissions on
`arn:aws:rds:${data.aws_region.current.name}:pg:default.*` are required
so that we can add the RDS instance to the parameter group's list of RDS
instance members.

Permissions on `aws_db_parameter_group.custom_parameters.arn` are
required to allow the IAM user to switch the RDS instance **back** to
the original parameter group (they shouldn't need to do this, during an
upgrade, but it allows them to roll back the first step, just in case).