Correct var. to `infrastructure_support`
When we create an environment using the CLI tool, the variable
`infrastructure_support` is defined in `variables.tf`

This change corrects the template to use the correct variable name.
Prior to this, the following sequence of commands will not work:

```
cloud-platform environment create
cloud-platform environment rds create
```

...because the pipeline will throw this error:

```
Error: Reference to undeclared input variable

  on rds.tf line 18, in module "rds":
  18:   infrastructure-support = var.infrastructure-support

An input variable with the name "infrastructure-support" has not been
declared. Did you mean "infrastructure_support"?
```