Allow maintenance window to be configured
We'd like to make our maintenance window more explicit so it matches the
time when we have the least number of users of our tool. This change
allows the window to be set in much the same way that the
`backup_window` attribute can be set.

See
http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_UpgradeDBInstance.Maintenance.html#AdjustingTheMaintenanceWindow
for more information about RDS maintenance windows.