Add permissions for accessing database logs (#120)
* Sort IAM policy actions

To make it easier to find actions in the list.

* Add permissions for accessing RDS logs

This allows users of this Terraform module to use the AWS credentials to
get a list of logs and download the contents of the logs. This is
necessary as users within MoJ don't currently have access to the AWS
console.